streamline module creation :change:
The way it was I frequently had problems on the first import of a
module, I played around, added time.sleep between the creation and the
import but the same import still failed, although i even could in the
meantime import the just generated module from another process fine.
No it always uses the same code as if no gen_dir was defined to import
the module on the first creation, however now one more test of the leak
test is failing on my test machine, don't know why.